Split primitives out into their own modules
