+++ crates/opencascade/src/primitives/edge.rs
@@ -1,0 +1,107 @@
+use crate::primitives::{make_axis_2, make_point};
+use cxx::UniquePtr;
+use glam::{dvec3, DVec3};
+use opencascade_sys::ffi::{self};
+
+pub struct Edge {
+    pub(crate) inner: UniquePtr<ffi::TopoDS_Edge>,
+}
+
+impl AsRef<Edge> for Edge {
+    fn as_ref(&self) -> &Edge {
+        self
+    }
+}
+
+impl Edge {
+    pub fn segment(p1: DVec3, p2: DVec3) -> Self {
+        let mut make_edge =
+            ffi::BRepBuilderAPI_MakeEdge_gp_Pnt_gp_Pnt(&make_point(p1), &make_point(p2));
+        let edge = make_edge.pin_mut().Edge();
+        let inner = ffi::TopoDS_Edge_to_owned(edge);
+
+        Self { inner }
+    }
+
+    pub fn circle(center: DVec3, normal: DVec3, radius: f64) -> Self {
+        let axis = make_axis_2(center, normal);
+
+        let make_circle = ffi::gp_Circ_ctor(&axis, radius);
+
+        let mut make_edge = ffi::BRepBuilderAPI_MakeEdge_circle(&make_circle);
+
+        let edge = make_edge.pin_mut().Edge();
+        let inner = ffi::TopoDS_Edge_to_owned(edge);
+
+        Self { inner }
+    }
+
+    pub fn ellipse() {}
+
+    pub fn spline() {}
+
+    pub fn arc(p1: DVec3, p2: DVec3, p3: DVec3) -> Self {
+        let make_arc = ffi::GC_MakeArcOfCircle_point_point_point(
+            &make_point(p1),
+            &make_point(p2),
+            &make_point(p3),
+        );
+
+        let mut make_edge = ffi::BRepBuilderAPI_MakeEdge_HandleGeomCurve(
+            &ffi::new_HandleGeomCurve_from_HandleGeom_TrimmedCurve(&ffi::GC_MakeArcOfCircle_Value(
+                &make_arc,
+            )),
+        );
+
+        let edge = make_edge.pin_mut().Edge();
+        let inner = ffi::TopoDS_Edge_to_owned(edge);
+
+        Self { inner }
+    }
+
+    pub fn start_point(&self) -> DVec3 {
+        let curve = ffi::BRepAdaptor_Curve_ctor(&self.inner);
+        let start_param = curve.FirstParameter();
+        let point = ffi::BRepAdaptor_Curve_value(&curve, start_param);
+
+        dvec3(point.X(), point.Y(), point.Z())
+    }
+
+    pub fn end_point(&self) -> DVec3 {
+        let curve = ffi::BRepAdaptor_Curve_ctor(&self.inner);
+        let last_param = curve.LastParameter();
+        let point = ffi::BRepAdaptor_Curve_value(&curve, last_param);
+
+        dvec3(point.X(), point.Y(), point.Z())
+    }
+
+    pub fn approximation_segments(&self) -> ApproximationSegmentIterator {
+        let adaptor_curve = ffi::BRepAdaptor_Curve_ctor(&self.inner);
+        let approximator = ffi::GCPnts_TangentialDeflection_ctor(&adaptor_curve, 0.1, 0.1);
+
+        ApproximationSegmentIterator { count: 1, approximator }
+    }
+
+    pub fn tangent_arc(_p1: DVec3, _tangent: DVec3, _p3: DVec3) {}
+}
+
+pub struct ApproximationSegmentIterator {
+    count: usize,
+    approximator: UniquePtr<ffi::GCPnts_TangentialDeflection>,
+}
+
+impl Iterator for ApproximationSegmentIterator {
+    type Item = DVec3;
+
+    fn next(&mut self) -> Option<Self::Item> {
+        if self.count <= self.approximator.NbPoints() as usize {
+            let point =
+                ffi::GCPnts_TangentialDeflection_Value(&self.approximator, self.count as i32);
+
+            self.count += 1;
+            Some(dvec3(point.X(), point.Y(), point.Z()))
+        } else {
+            None
+        }
+    }
+}
