+++ crates/opencascade/src/primitives/shell.rs
@@ -1,0 +1,12 @@
+use cxx::UniquePtr;
+use opencascade_sys::ffi::{self};
+
+pub struct Shell {
+    pub(crate) _inner: UniquePtr<ffi::TopoDS_Shell>,
+}
+
+impl AsRef<Shell> for Shell {
+    fn as_ref(&self) -> &Shell {
+        self
+    }
+}
